Create .gitignore
Ignore pycache files. Got the idea because I saw a pycache file once and it annoyed me to no end. I hate pycache files. I hate them a lot. I looked in the 2021-2022 code and saw that there was a nifty .gitignore thing you can do, so I did it.